+++ index.js
@@ -9,7 +9,8 @@
 // It's easier just to leave it here.
 import App from "./app/app.tsx"
 import { AppRegistry } from "react-native"
-
+import { NativeModules } from 'react-native'
+console.log(NativeModules.CargoLinkMobile)
 /**
  * This needs to match what's found in your app_delegate.m and MainActivity.java.
  */