+++ CHANGELOG.md
@@ -2,6 +2,32 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.1.17](https://git-codecommit.ap-southeast-1.amazonaws.com///compare/v1.1.15...v1.1.17) (2021-05-19)
+
+
+### Features
+
+* add version into react-native-maps ([86727f7](https://git-codecommit.ap-southeast-1.amazonaws.com///commit/86727f7b9be9f98999544ceb6a24b868eaba2681))
+* show button book job when status createdFrom = 1 (app) ([cd0e65c](https://git-codecommit.ap-southeast-1.amazonaws.com///commit/cd0e65c10c9d677d7edd3bcdccab988b8d6cd4b1))
+
+
+### Bug Fixes
+
+* **fastlane:** conventional changelog ([014b21c](https://git-codecommit.ap-southeast-1.amazonaws.com///commit/014b21cf8c3e8055e8ea9da588aa8f9dd778beca))
+
+### [1.1.16](https://git-codecommit.ap-southeast-1.amazonaws.com///compare/v1.1.15...v1.1.16) (2021-05-19)
+
+
+### Features
+
+* add version into react-native-maps ([86727f7](https://git-codecommit.ap-southeast-1.amazonaws.com///commit/86727f7b9be9f98999544ceb6a24b868eaba2681))
+* show button book job when status createdFrom = 1 (app) ([cd0e65c](https://git-codecommit.ap-southeast-1.amazonaws.com///commit/cd0e65c10c9d677d7edd3bcdccab988b8d6cd4b1))
+
+
+### Bug Fixes
+
+* **fastlane:** conventional changelog ([014b21c](https://git-codecommit.ap-southeast-1.amazonaws.com///commit/014b21cf8c3e8055e8ea9da588aa8f9dd778beca))
+
 ### [1.1.15](https://git-codecommit.ap-southeast-1.amazonaws.com///compare/v1.1.14...v1.1.15) (2021-05-10)
 
 * bug fixes: crash in my job screen